backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
openai
chromadb
tiktoken
python-dateutil
pytz
python-multipart
